qtility-standard: remove hiding of `map`
Probably not worth subverting the expectations that people have,
removing `id` quickly turned out to be a mistake, so I'm guessing this
one would be too.

I think naming `fmap` `map` would overall be a much better idea, but I
think it'd have to be an ecosystem change. I'd hoped to make that change
in a smaller context for `Qtility.Standard` but I think it's easier to
just teach and move forward with `fmap` & `map` as they are.