feat: throw in `getSecret...` when secret value can't be found
This should be an exceptional case, in my opinion; if you're fetching
a secret it's most likely something you expect to be there and it should
be a rare case that it is not, and something you'd only want to guard
against in cases where this is embedded deep in a system, i.e. not in
initialization where you're more likely to call this.